Create stubs for registration views
This commit adds a few basic views for the registration forms and view-
ing responses:
- Institutions + Coach
- Teams + Speakers
- Adjudicators

When an institution is created, the coach will have a URL and access to
forms to register participants on their behalf. Also when a team is
created, speakers can register to that team (through an invitation TBI).